move formatting utils and savable api object to independent modules
