+++ featurebyte/common/utils.py
@@ -3,27 +3,21 @@ Utility functions for API Objects
 """
 from __future__ import annotations
 
-from typing import Any, Dict, Iterator, List, Optional, Union
+from typing import Any, Iterator, List, Optional, Union
 
-import copy
 import functools
-import re
 from datetime import datetime
 from decimal import Decimal
 from importlib import metadata as importlib_metadata
 from io import BytesIO
 from pathlib import Path
-from xml.dom.minidom import Document, Element, getDOMImplementation
 
 import numpy as np
 import pandas as pd
 import pyarrow as pa
 import pyarrow.parquet as pq
-import pygments
 from alive_progress import alive_bar
-from bson import ObjectId
 from dateutil import parser
-from pygments.formatters.html import HtmlFormatter
 from requests import Response
 
 from featurebyte.common.env_util import get_alive_bar_additional_params
@@ -406,150 +400,3 @@ def convert_to_list_of_strings(value: Optional[Union[str, List[str]]]) -> List[s
     if isinstance(value, list):
         output = value
     return output
-
-
-class CodeStr(str):
-    """
-    Code string content that can be displayed in markdown format
-    """
-
-    def _repr_html_(self) -> str:
-        lexer = pygments.lexers.get_lexer_by_name("python")
-        highlighted_code = pygments.highlight(
-            str(self).strip(),
-            lexer=lexer,
-            formatter=HtmlFormatter(noclasses=True, nobackground=True),
-        )
-        return (
-            '<div style="margin:30px; padding: 20px; border:1px solid #aaa">'
-            f"{highlighted_code}</div>"
-        )
-
-
-class InfoDict(Dict[str, Any]):
-    """
-    Featurebyte asset information dictionary that can be displayed in HTML
-    """
-
-    def __init__(self, data: Dict[str, Any]) -> None:
-        self.class_name = data.pop("class_name", "Unknown")
-        super().__init__(data)
-
-    def _repr_html_(self) -> str:
-        def _set_element_style(elem: Element, style: Dict[str, Any]) -> None:
-            """
-            Set style of dom element
-
-            Parameters
-            ----------
-            elem: Element
-                Element to set style for.
-            style: Dict[str, Any]
-                Style dictionary
-            """
-            elem.setAttribute("style", ";".join(f"{key}:{value}" for key, value in style.items()))
-
-        def _populate_html_elem(
-            data: Dict[str, Any], doc: Document, elem: Element, html_content: Dict[str, str]
-        ) -> None:
-            """
-            Populate html document with data dict
-
-            Parameters
-            ----------
-            data: Dict[str, str]
-                Data dictionary
-            doc: Document
-                HTML document
-            elem: Element
-                HTML element to populate into
-            html_content: Dict[str, str]
-                Dictionary referencing HTML content in the document
-            """
-            # create html table
-            table = doc.createElement("table")
-            _set_element_style(table, {"width": "100%", "padding": 0, "margin": 0})
-
-            for key, value in data.items():
-                # add row to table
-                row = doc.createElement("tr")
-                table.appendChild(row)
-
-                # populate key column
-                key_column = doc.createElement("td")
-                _set_element_style(
-                    key_column,
-                    {
-                        "font-weight": "bold",
-                        "vertical-align": "top",
-                        "width": "200px",
-                        "over-flow": "overflow-wrap",
-                        "word-break": "break-all",
-                    },
-                )
-                key_column.appendChild(doc.createTextNode(key))
-                row.appendChild(key_column)
-
-                # populate value column
-                if isinstance(value, dict):
-                    # process dictionaries recursively
-                    value_elem = doc.createElement("div")
-                    _set_element_style(value_elem, {"border": "0", "padding": "0", "margin": "0"})
-                    _populate_html_elem(
-                        data=value, doc=doc, elem=value_elem, html_content=html_content
-                    )
-                elif isinstance(value, list) and isinstance(value[0], dict):
-                    # list of dictionaries
-                    value_elem = doc.createElement("div")
-                    _set_element_style(value_elem, {"border": "0", "padding": "0", "margin": "0"})
-                    df_key = str(ObjectId())
-                    html_content[df_key] = pd.DataFrame(value).to_html()
-                    value_elem.appendChild(doc.createTextNode(f"{{{df_key}}}"))
-                else:
-                    # nicer datetime formatting
-                    try:
-                        value = datetime.fromisoformat(value).strftime("%Y-%m-%d %H:%M:%S")
-                    except (ValueError, TypeError):
-                        pass
-                    value_elem = doc.createTextNode(str(value))
-
-                val_column = doc.createElement("td")
-                _set_element_style(val_column, {"width": "80%", "text-align": "left"})
-                val_column.appendChild(value_elem)
-                row.appendChild(val_column)
-
-            elem.appendChild(table)
-
-        # create html document
-        impl = getDOMImplementation()
-        assert impl
-        doc_type = impl.createDocumentType(
-            "html",
-            "-//W3C//DTD XHTML 1.0 Strict//EN",
-            "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd",
-        )
-        doc = impl.createDocument("http://www.w3.org/1999/xhtml", "html", doc_type)
-        doc_elem = doc.documentElement
-
-        # add title for the table
-        data = copy.deepcopy(self)
-        class_name = re.sub(r"([A-Z]+[a-z]+)", r" \1", self.class_name).strip()
-        title_div = doc.createElement("div")
-        title_div.appendChild(doc.createTextNode(class_name))
-        _set_element_style(
-            elem=title_div,
-            style={
-                "font-weight": "bold",
-                "text-align": "center",
-                "border-bottom": "1px solid black",
-                "width": "100%",
-                "padding-top": "20px",
-            },
-        )
-        doc_elem.appendChild(title_div)
-
-        # add info table
-        html_content: Dict[str, str] = {}
-        _populate_html_elem(data=data, doc=doc, elem=doc_elem, html_content=html_content)
-        html = str(doc.toprettyxml())
-        return html.format(**html_content)
